refactor: restructure to clean architecture
