+++ .gitignore
@@ -1,5 +1,6 @@
-obj/*
+**/obj/*
 .idea/*
-bin/*
+**/bin/*
 orders.db
-*.user
+**/*.user
+.trash/*
