--- Data/IDbConnectionFactory.cs
@@ -1,9 +1,0 @@
-using System.Data;
-
-namespace LegacyOrderService.Data
-{
-    public interface IDbConnectionFactory
-    {
-        IDbConnection CreateConnection();
-    }
-}
