--- Program.cs
@@ -1,28 +1,0 @@
-
-
-using LegacyOrderService.Data;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Hosting;
-using Microsoft.Extensions.Configuration;
-using LegacyOrderService.Application;
-using Serilog;
-
-Log.Logger = new LoggerConfiguration()
-    .WriteTo.Console()
-    .WriteTo.File("app.log", rollingInterval: RollingInterval.Day)
-    .CreateLogger();
-
-var host = Host.CreateDefaultBuilder(args)
-    .UseSerilog()
-    .ConfigureServices((ctx, services) =>
-    {
-        var connectionString = ctx.Configuration.GetConnectionString("Default");
-        services.AddSingleton<IDbConnectionFactory>(new SqliteConnectionFactory(connectionString));
-        services.AddScoped<IOrderRepository, OrderRepository>();
-        services.AddScoped<IProductRepository, ProductRepository>();
-        services.AddScoped<IOrderService, OrderService>();
-        services.AddScoped<App>();
-    })
-    .Build();
-
-host.Services.GetRequiredService<App>().Run(args);
