--- appsettings.json
@@ -1,5 +1,0 @@
-{
-  "ConnectionStrings": {
-    "Default": "Data Source=orders.db"
-  }
-}
