+++ src/OrderService.Domain/Interfaces/IProductRepository.cs
@@ -1,4 +1,4 @@
-namespace LegacyOrderService.Data
+namespace OrderService.Domain.Interfaces
 {
     public interface IProductRepository
     {
